Remove hard dependency on skeleton-i18n
Move the package to suggests instead and handle cases where it is not
installed.

+++ lib/Skeleton/Object/Delete.php
@@ -17,7 +17,6 @@ trait Delete {
 	 * @access public
 	 */
 	public function delete() {
-
 		if (method_exists($this, 'trait_child_delete') and is_callable([$this, 'trait_child_delete'])) {
 			$this->trait_child_delete();
 		}
@@ -26,6 +25,10 @@ trait Delete {
 		$db = self::trait_get_database();
 
 		if (isset(self::$object_text_fields)) {
+			if (!class_exists('\Skeleton\I18n\Object\Text')) {
+				throw new \Exception('Skeleton package "skeleton-i18n" needs to be installed to use object text');
+			}
+
 			$object_texts = \Skeleton\I18n\Object\Text::get_by_object($this);
 			foreach ($object_texts as $object_text) {
 				$object_text->delete();
